Added text generation model and pre-trained model
This commit contains the textgenerationnietsche.py and texttestniet.py scripts, as well as the textgennietsche.h5 pre-trained model. The text generation model is trained using the textgenerationnietsche.py script. It imports Nietzsche's text, transforms the data into character sequences, and creates an LSTM model for text generation. A categorical cross-entropy loss function and the RMSprop optimizer are used to train the model with the processed text data. Textgennietsche.h5 is then stored as the trained model. The texttestniet.py script loads a previously-trained LSTM model and a user-supplied seed text. It generates new text based on the source text while imitating Nietzsche's writing style. The textgennietsche.h5 file is the pre-trained model that enables users to generate text without requiring additional training.